Audit-Webapp/requirements.txt
Flask==1.1.1
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.19.0
requests==2.22.0
urllib3==1.25.3

-e git+git://github.com/smart-on-fhir/client-py.git@feature/r4#egg=fhirclient